backend/requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# 向量数据库
chromadb==0.4.22

# AI/ML
openai==1.3.7
numpy==1.24.3
scikit-learn==1.3.2

# 文档处理
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
markdown==3.5.1

# 工具库
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0